Adding PoolTableList class.
Adds players to a waiting list and adds them to tables when available.
Also adds tables for players to play on.
Possesses function to allow the maximum amount of players to be sent to
all available players.
Checks whether the table is full or not before attempting to send
players to them.  Also checks if all tables are full before trying to
send the next player in line to any table.